Modify data types in messages: browsers do not support 64bit integers so keep everything at 32 bits.

+++ src/connector/instruction/WebXConnectInstruction.h
@@ -6,7 +6,7 @@
 class WebXConnectInstruction : public WebXInstruction {
 
 public:
-    WebXConnectInstruction(unsigned long id) :
+    WebXConnectInstruction(uint32_t id) :
         WebXInstruction(Type::Connect, id) {
     }
